Add custom exceptions, adjust patterns for sync/async, modify tests for httpx
