+++ README.md
@@ -19,7 +19,7 @@ This client is designed to be simple and easy to use. Here's a quick example of 
 keywords by ASIN:
 
 ```python
-from junglescout.client import Client
+from junglescout import Client
 from junglescout.models.parameters import Marketplace, ApiType, FilterOptions, Sort
 
 API_KEY_NAME = "api_key_name"
