+++ pyproject.toml
@@ -33,6 +33,7 @@ sphinx-autodoc-typehints = "^2.0.0"
 sphinx-rtd-theme = "^2.0.0"
 python-semantic-release = "^9.1.1"
 autodoc-pydantic = "^2.1.0"
+respx = "^0.21.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -41,7 +42,6 @@ build-backend = "poetry.core.masonry.api"
 [tool.black]
 line-length = 120
 target-version = ['py39']
-skip-numeric-underscore-normalization = true
 
 [tool.isort]
 profile = "black"
