+++ src/junglescout/__init__.py
@@ -3,8 +3,10 @@
 This package contains the Jungle Scout Client package.
 """
 
-__all__ = ["Client"]
+__all__ = ["Client", "ClientAsync", "ClientSync"]
 __version__ = "0.2.2"
 
-from .client import Client
-from .client_async import AsyncClient
+from .client_async import ClientAsync
+from .client_sync import ClientSync
+
+Client = ClientSync
