+++ tests/client_sync/test_client_instantiate.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from junglescout.client import Client
+from junglescout import Client
 from junglescout.models.parameters import Marketplace
