+++ tests/factories/keywords_by_keyword_factory.py
@@ -1,3 +1,5 @@
+from typing import Any
+
 import factory
 from faker import Faker
 
@@ -45,5 +47,5 @@ class KeywordsByKeywordResponseFactory(factory.DictFactory):
     meta = factory.LazyAttribute(lambda o: {"total_items": o.total_items})
 
 
-def generate_keywords_by_keyword_responses(total_items: int = 1):
+def generate_keywords_by_keyword_responses(total_items: int = 1) -> dict[str, Any]:
     return KeywordsByKeywordResponseFactory(total_items=total_items)
