+++ tests/integration/test_historical_search_volume.py
@@ -1,9 +1,10 @@
 from datetime import datetime
 
 import pytest
-from requests.exceptions import HTTPError
+from httpx import HTTPStatusError
 
-from junglescout.client import Client
+from junglescout import Client
+from junglescout.exceptions import JungleScoutHTTPError
 from junglescout.models.parameters import Marketplace
 
 
@@ -24,9 +25,10 @@ def test_historical_search_volume(api_keys):
 def test_historical_search_volume_with_too_large_of_range(api_keys):
     keyword = "yoga"
     client = Client(**api_keys, marketplace=Marketplace.US)
-    with pytest.raises(HTTPError) as exc_info:
+    with pytest.raises(JungleScoutHTTPError) as exc_info:
         client.historical_search_volume(keyword=keyword, start_date="2023-01-01", end_date="2024-04-30")
-    assert exc_info.value.response.json() == {
+    assert isinstance(exc_info.value.httpx_exception, HTTPStatusError)
+    assert exc_info.value.httpx_exception.response.json() == {
         "errors": [
             {
                 "title": "Invalid parameter value for: 'start_date'",
