+++ tests/integration/test_product_database.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import pytest
 
-from junglescout.client import Client
+from junglescout import Client
 from junglescout.models.parameters import (
     Marketplace,
     ProductSort,
